current php files

+++ processartists.php
@@ -5,11 +5,11 @@ $debugOn = true;
 <html>
 <head>
 <meta charset="UTF-8">
+    <link href="styles.css" rel="stylesheet" type="text/css">
 <title>Artists Records</title>
 </head>
 
 <body>
-<h1>Results</h1>
 <?php
 
 function upload_photo($ext,$filename){
@@ -38,18 +38,13 @@ echo "<p>Your image upload was successful!\n</p>";
 }
 }
 
-//echo "<h2>Form Data</h2>";
-//echo "<pre>";
-//print_r($_REQUEST); // a useful debugging function to see everything in an array, best inside a <pre> element
-//echo "</pre>";
-
 // execute the appropriate query based on which submit button (insert, delete or update) was clicked
-if (isset($_POST))
+if ($_POST['submit'])
 {
 	$errorMessage="";
     if(empty($_POST['name']))
     {
-        $errorMessage.="<li>Please enter the artists name</li>";
+        $errorMessage.="Please enter the artist's name";
     }
     else{
             //check if artist already in database
@@ -57,16 +52,15 @@ if (isset($_POST))
     $rows = $result->fetchall(PDO::FETCH_ASSOC);
 	foreach ($rows as $row){
     if ($row[name]==$_REQUEST[name] AND $_REQUEST['submit']=="Add Entry"){
-            $errorMessage.="<li>Artist already exists, please edit details below</li>";
+            $errorMessage.="Artist already exists, please edit details below";
         }
      }
     }
 
     if(!empty($errorMessage))
     {
-        echo("<p>There was an error with your form:
-        <ul>" . $errorMessage . "</ul>\n</p>");
-		include "artists.php";
+        include "artists.php";
+		
 	 }
 else{
 if ($_REQUEST['submit'] == "Add Entry")
@@ -77,7 +71,6 @@ upload_photo($ext, $filename);
 
     
 	$sql = "INSERT INTO artists (name, shortBio, fullBio,filename) VALUES ('$_REQUEST[name]', '$_REQUEST[shortBio]', '$_REQUEST[fullBio]', '$filename')";
-	//echo "<p>Query: " . $sql . "</p>\n";
 	if ($dbh->exec($sql)){
 		echo "$_REQUEST[name] has been added.";
 		}
@@ -92,41 +85,36 @@ upload_photo($ext, $filename);
 
 
 		$sq="INSERT INTO Genre (id,category) VALUES($id[id],'$val')";
-		//echo "<p>Query: " . $sq . "</p>\n";
-	if ($dbh->exec($sq)){
-		echo "<p><strong>Inserted $_REQUEST[name]"."\n</strong></p>";
-		}
-	else{
-		echo "<p>Not inserted\n</p>"; // in case it didn't work - e.g. if database is not writeable	
-	}
+	   $dbh->exec($sq);
 	}
 }
 
 
-else if ($_REQUEST['submit'] == "Delete Entry")
+else if ($_REQUEST['submit'] == "Delete")
 {
-    if(unlink(__DIR__.'/uploads/'.$_REQUEST[filename]))
-        echo "image deleted";
+    unlink(__DIR__.'/uploads/'.$_REQUEST[filename]);
 
+    $sql="DELETE FROM Genre WHERE id ='$_REQUEST[id]'";
+    $dbh->exec($sql);
     
 	$sql = "DELETE FROM artists WHERE id = '$_REQUEST[id]'";
-	echo "<p>Query: " . $sql . "</p>\n<p><strong>"; 
-	if ($dbh->exec($sql) AND $dbh->exec($sq))
+	if ($dbh->exec($sql))
 		echo "Deleted $_REQUEST[name]";
 	else
         echo "Not deleted";
     
 }
 
-else if ($_REQUEST['submit'] == "Update Entry")
+else if ($_REQUEST['submit'] == "Update")
 {
+if (is_uploaded_file($_FILES['photo']['tmp_name'])) {
 $filepath = __DIR__.'/uploads/'.$_REQUEST[filename];
 unlink($filepath);
-
-    $ext = substr($_FILES['photo']['name'], strpos($_FILES['photo']['name'],'.'), strlen($_FILES['photo']['name'])-1);
-    $filename = $_FILES['photo']['name'];
-    upload_photo($ext, $filename);
-
+$ext = substr($_FILES['photo']['name'], strpos($_FILES['photo']['name'],'.'), strlen($_FILES['photo']['name'])-1);
+$filename = $_FILES['photo']['name'];
+upload_photo($ext, $filename);
+}
+else {$filename= $_REQUEST['filename'];}
 
 	$sql = "UPDATE artists SET name = '$_REQUEST[name]', shortBio = '$_REQUEST[shortBio]', fullBio = '$_REQUEST[fullBio]', filename='$filename' WHERE id = '$_REQUEST[id]'";
 	//echo "<p>Query: " . $sql . "</p>\n<p><strong>";
@@ -136,53 +124,44 @@ unlink($filepath);
 		echo "Not updated";
 
 	$sq = "DELETE FROM Genre WHERE id = '$_REQUEST[id]'";
-	if ($dbh->exec($sq)){
-	//echo "<p>Query: " . $sq . "</p>\n<p><strong>";
-	}
+	$dbh->exec($sq);
 	foreach ($_POST['category'] as $val){
           $sq="SELECT id FROM artists WHERE name='$_REQUEST[name]'";
           $result=($dbh->query($sq));
           $id=$result->fetch();
 
   	$sql="INSERT INTO Genre (id,category) VALUES($id[id],'$val')";
-  	echo "<p>Query: " . $sql . "</p>\n<p><strong>";}
-
-    	if ($dbh->exec($sql))
-    		echo "Updated $_REQUEST[name]";
-    	else
-    		echo "Not updated";
-
+  	 }
+    	$dbh->exec($sql);
 }
 
 else {
 	echo "This page did not come from a valid form submission.<br />\n";
+    }
 }
-
+}
+    
 // Basic select and display all contents from table 
-echo "<h2>Artists</h2>\n";
-$sql = "SELECT name,shortBio FROM artists";
+$sql = "SELECT id,name,shortBio,filename FROM artists";
 $result = $dbh->query($sql);
 
-if ($debugOn) {
-	echo "<pre>";
-	$rows = $result->fetchall(PDO::FETCH_ASSOC);
-    echo "$rows[name]=".$rows[name];
-    echo "$rows[1][name]".$rows[1][name];
-	echo count($rows) . " records in table<br />\n";
-	print_r($rows);
-    $img="uploads/".$filename;
-    echo '<img src="'.$img.'">';
-	echo "</pre>";
-//	echo "<br />\n";
+while ($row=$result->fetch()){
+    echo"<a href='artistpage.php?id=$row[id]'><div id='artist'>";
+    if($row['filename']){
+    echo "<img id='smallimage' src=\"uploads/$row[filename]\">";
+    }
+    echo "<div><h1>".$row['name']."</h1>";
+    echo "<p>".$row['shortBio']."</p><br></div></div></a>";
+    
 }
-}
-}
+    
+
 
 // close the database connection 
 $dbh = null;
 ?>
 
-<p><a href="artists.php">Return to database test page</a></p>
+<p><a href="artist.php">Return to database test page</a></p>
 
 </body>
 </html>
